Init persistent subscription on start  - Normalize config  - Cleanup interfaces
